backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0
django-filter>=24.0,<25.0
